refactor: Rename utils to helpers and update imports for consistency
