+++ src/helpers/handCalc/patterns.ts
