+++ src/helpers/randomDice.ts
