+++ src/pages/Home/Game/Modal/components/Dices/icons/Balas.tsx
