+++ src/pages/Home/Game/Modal/components/Dices/icons/Cuadras.tsx
