+++ src/pages/Home/Game/Modal/components/Dices/icons/Quinas.tsx
