+++ src/pages/Home/Game/Modal/components/Dices/icons/Senas.tsx
