+++ src/pages/Home/Game/Modal/components/Dices/icons/Stripe.tsx
