+++ src/pages/Home/Game/Modal/components/Dices/icons/Text.tsx
