+++ src/pages/Home/Game/Modal/components/Dices/icons/Tontos.tsx
