+++ src/pages/Home/Game/Modal/components/Dices/icons/Trenes.tsx
