+++ src/pages/Home/Game/Modal/components/Dices/icons/index.ts
