+++ src/pages/Home/Game/Modal/components/Dices/index.tsx
