+++ src/pages/Home/Game/Modal/components/Status.tsx
